Updated the kill to avoid running with error if not started

+++ kill.sh
@@ -1,1 +1,8 @@
-kill $(ps -efw | grep connector.py | grep -v grep | awk '{print $2}')
+myvar=$(ps -efw | grep connector.py | grep -v grep | awk '{print $2}')
+if test -z "$myvar"
+then
+  echo "empty"
+else
+  echo "not empty"
+  kill $(ps -efw | grep connector.py | grep -v grep | awk '{print $2}')
+fi
